Started on separation of main class
